requirements.txt
Flask==2.3.3
paramiko==3.3.1
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.103.0
faker==19.6.2
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
